Removed all .Java gateway commands and moved into -Data commands.

+++ src/main/java/com/github/princesslana/eriscasper/gateway/Payload.java
@@ -35,5 +35,5 @@ public abstract class Payload {
 
   @Value.Immutable
   @Wrapped
-  public static interface SequenceNumberWrapper extends Wrapper<Integer> {}
+  public static interface SequenceNumberWrapper extends Wrapper<Long> {}
 }
